+++ src/main/java/com/example/developmentquiz/domain/Product.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class Product {
     private String imgURL;
     private String productName;
-    private int unitPrice;
+    private float unitPrice;
     private String unitType;
 }
